Convert NewRuntimeModal to be a function component (#2590)
