+++ src/components/AnalysisModal.js
@@ -2,11 +2,11 @@ import _ from 'lodash/fp'
 import { Fragment, useState } from 'react'
 import { div, h, h2, hr, img, span } from 'react-hyperscript-helpers'
 import { ButtonPrimary, IdContainer, Select, spinnerOverlay, WarningTitle } from 'src/components/common'
+import { ComputeModalBase } from 'src/components/ComputeModal'
 import Dropzone from 'src/components/Dropzone'
+import { GalaxyModalBase } from 'src/components/GalaxyModal'
 import { icon } from 'src/components/icons'
 import ModalDrawer from 'src/components/ModalDrawer'
-import { NewGalaxyModalBase } from 'src/components/NewGalaxyModal'
-import { NewRuntimeModalBase } from 'src/components/NewRuntimeModal'
 import { analysisNameInput, analysisNameValidator, getDisplayName, getTool, notebookData, tools } from 'src/components/notebook-utils'
 import TitleBar from 'src/components/TitleBar'
 import galaxyLogo from 'src/images/galaxy-logo.png'
@@ -22,19 +22,21 @@ import * as Utils from 'src/libs/utils'
 import validate from 'validate.js'
 
 
-const titleId = 'new-analysis-modal-title'
-const newAnalysisMode = Symbol('new-artifact')
-const newEnvironmentMode = Symbol('new-environment')
+const titleId = 'analysis-modal-title'
+const analysisMode = Symbol('artifact')
+const environmentMode = Symbol('environment')
 
-export const NewAnalysisModal = Utils.withDisplayName('NewAnalysisModal')(
-  ({ isOpen, onDismiss, onSuccess, uploadFiles, openUploader, runtimes, apps, galaxyDataDisks, refreshRuntimes, refreshApps, refreshAnalyses, analyses, workspace, persistentDisks, workspace: { workspace: { namespace, bucketName, name: workspaceName } } }) => {
+export const AnalysisModal = Utils.withDisplayName('AnalysisModal')(
+  ({
+    isOpen, onDismiss, onSuccess, uploadFiles, openUploader, runtimes, apps, galaxyDataDisks, refreshRuntimes, refreshApps, refreshAnalyses,
+    analyses, workspace, persistentDisks, workspace: { workspace: { namespace, bucketName, name: workspaceName } }
+  }) => {
     const [viewMode, setViewMode] = useState(undefined)
     const [busy, setBusy] = useState()
     const [notebookKernel, setNotebookKernel] = useState('python3')
     const [analysisName, setAnalysisName] = useState('')
     const prevAnalysisName = Utils.usePrevious(analysisName)
     const [currentTool, setCurrentTool] = useState(undefined)
-
 
     const currentRuntime = getCurrentRuntime(runtimes)
     const currentRuntimeTool = currentRuntime?.labels?.tool
@@ -47,33 +49,36 @@ export const NewAnalysisModal = Utils.withDisplayName('NewAnalysisModal')(
       setNotebookKernel('python3')
     }
 
-    // The intended flow is to call this without a viewmode, and have it intelligently figure out the next step for you
-    // Passing a viewmode is a way to force your next modal
+    /**
+     * The intended flow is to call this without a viewMode, and have it intelligently figure out the next
+     * step for you. Passing a viewMode is a way to force your next modal.
+     */
     const enterNextViewMode = (currentTool, baseViewMode = viewMode) => {
       const doesCloudEnvForToolExist = currentRuntimeTool === currentTool || (currentApp && currentTool === tools.galaxy.label)
 
       Utils.switchCase(baseViewMode,
-        [newAnalysisMode, () => Utils.cond(
+        [analysisMode, () => Utils.cond(
           [doesCloudEnvForToolExist, () => {
             resetView()
             onSuccess()
           }],
-          [!doesCloudEnvForToolExist && currentRuntime && isRuntimeDeletable(currentRuntime), () => setViewMode(newEnvironmentMode)],
-          [!doesCloudEnvForToolExist && !currentRuntime, () => setViewMode(newEnvironmentMode)],
+          [!doesCloudEnvForToolExist && currentRuntime && isRuntimeDeletable(currentRuntime), () => setViewMode(environmentMode)],
+          [!doesCloudEnvForToolExist && !currentRuntime, () => setViewMode(environmentMode)],
           [!doesCloudEnvForToolExist && currentRuntime && !isRuntimeDeletable(currentRuntime), () => {
             resetView()
             onSuccess()
           }]
         )],
-        [newEnvironmentMode, () => {
+        [environmentMode, () => {
           resetView()
           onSuccess()
         }],
         [Utils.DEFAULT, () => Utils.cond(
-          [currentTool === tools.RStudio.label || currentTool === tools.Jupyter.label, () => setViewMode(newAnalysisMode)],
-          [currentTool === tools.galaxy.label && !currentApp, () => setViewMode(newEnvironmentMode)],
+          [currentTool === tools.RStudio.label || currentTool === tools.Jupyter.label, () => setViewMode(analysisMode)],
+          [currentTool === tools.galaxy.label && !currentApp, () => setViewMode(environmentMode)],
           [currentTool === tools.galaxy.label && currentApp, () => {
-            console.error('This shouldn\'t be possible, as you aren\'t allowed to create a galaxy analysis when one exists; the button should be disabled.')
+            console.error(
+              'This shouldn\'t be possible, as you aren\'t allowed to create a galaxy analysis when one exists; the button should be disabled.')
             resetView()
           }]
         )]
@@ -81,17 +86,17 @@ export const NewAnalysisModal = Utils.withDisplayName('NewAnalysisModal')(
     }
 
     const getView = () => Utils.switchCase(viewMode,
-      [newAnalysisMode, renderCreateAnalysis],
-      [newEnvironmentMode, getNewEnvironmentView],
+      [analysisMode, renderCreateAnalysis],
+      [environmentMode, getEnvironmentView],
       [Utils.DEFAULT, renderSelectAnalysisBody])
 
-    const getNewEnvironmentView = () => Utils.switchCase(currentTool,
-      [tools.Jupyter.label, renderNewRuntimeModal],
-      [tools.RStudio.label, renderNewRuntimeModal],
-      [tools.galaxy.label, renderNewGalaxyModal]
+    const getEnvironmentView = () => Utils.switchCase(currentTool,
+      [tools.Jupyter.label, renderComputeModal],
+      [tools.RStudio.label, renderComputeModal],
+      [tools.galaxy.label, renderGalaxyModal]
     )
 
-    const renderNewRuntimeModal = () => h(NewRuntimeModalBase, {
+    const renderComputeModal = () => h(ComputeModalBase, {
       isOpen: currentTool === tools.Jupyter.label || currentTool === tools.RStudio.label,
       isAnalysisMode: true,
       workspace,
@@ -103,7 +108,7 @@ export const NewAnalysisModal = Utils.withDisplayName('NewAnalysisModal')(
         onDismiss()
       },
       onSuccess: _.flow(
-        withErrorReporting('Error creating runtime'),
+        withErrorReporting('Error creating compute'),
         Utils.withBusyState(setBusy)
       )(async () => {
         setViewMode(undefined)
@@ -112,7 +117,7 @@ export const NewAnalysisModal = Utils.withDisplayName('NewAnalysisModal')(
       })
     })
 
-    const renderNewGalaxyModal = () => h(NewGalaxyModalBase, {
+    const renderGalaxyModal = () => h(GalaxyModalBase, {
       isOpen: viewMode === tools.galaxy.label,
       isAnalysisMode: true,
       workspace,
@@ -133,17 +138,39 @@ export const NewAnalysisModal = Utils.withDisplayName('NewAnalysisModal')(
     })
 
     const styles = {
-      toolCard: { backgroundColor: 'white', borderRadius: 5, padding: '1rem', display: 'inline-block', verticalAlign: 'middle', marginBottom: '1rem', textAlign: 'center', width: '100%', height: 60 },
+      toolCard: {
+        backgroundColor: 'white', borderRadius: 5, padding: '1rem', display: 'inline-block', verticalAlign: 'middle', marginBottom: '1rem',
+        textAlign: 'center', width: '100%', height: 60
+      },
       image: { verticalAlign: 'middle', height: '100%', width: '40%' }
     }
 
-    const renderToolButtons = () => div({ style: { display: 'flex', alignItems: 'center', flexDirection: 'column', justifyContent: 'space-between' } }, [
-      div({ style: styles.toolCard, onClick: () => { setCurrentTool(tools.Jupyter.label); enterNextViewMode(tools.Jupyter.label) } }, [img({ src: jupyterLogoLong, style: _.merge(styles.image, { width: '30%' }) })]),
-      div({ style: styles.toolCard, onClick: () => { setCurrentTool(tools.RStudio.label); enterNextViewMode(tools.RStudio.label) } }, [img({ src: rstudioLogo, style: styles.image })]),
-      div({ style: { opacity: currentApp ? '0.5' : '1', ...styles.toolCard }, onClick: () => { setCurrentTool(tools.galaxy.label); enterNextViewMode(tools.galaxy.label) }, disabled: !currentApp, title: currentApp ? 'You already have a galaxy environment' : '' }, [img({ src: galaxyLogo, style: _.merge(styles.image, { width: '30%' }) })])
+    const renderToolButtons = () => div({
+      style: { display: 'flex', alignItems: 'center', flexDirection: 'column', justifyContent: 'space-between' }
+    }, [
+      div({
+        style: styles.toolCard, onClick: () => {
+          setCurrentTool(tools.Jupyter.label)
+          enterNextViewMode(tools.Jupyter.label)
+        }
+      }, [img({ src: jupyterLogoLong, style: _.merge(styles.image, { width: '30%' }) })]),
+      div({
+        style: styles.toolCard, onClick: () => {
+          setCurrentTool(tools.RStudio.label)
+          enterNextViewMode(tools.RStudio.label)
+        }
+      }, [img({ src: rstudioLogo, style: styles.image })]),
+      div({
+        style: { opacity: currentApp ? '0.5' : '1', ...styles.toolCard }, onClick: () => {
+          setCurrentTool(tools.galaxy.label)
+          enterNextViewMode(tools.galaxy.label)
+        }, disabled: !currentApp, title: currentApp ? 'You already have a galaxy environment' : ''
+      }, [img({ src: galaxyLogo, style: _.merge(styles.image, { width: '30%' }) })])
     ])
 
-    const renderSelectAnalysisBody = () => div({ style: { display: 'flex', flexDirection: 'column', flex: 1, padding: '0.5rem 1.5rem 1.5rem 1.5rem' } }, [
+    const renderSelectAnalysisBody = () => div({
+      style: { display: 'flex', flexDirection: 'column', flex: 1, padding: '0.5rem 1.5rem 1.5rem 1.5rem' }
+    }, [
       renderToolButtons(),
       h(Dropzone, {
         accept: `.${tools.Jupyter.ext}, .${tools.RStudio.ext}`,
@@ -154,7 +181,9 @@ export const NewAnalysisModal = Utils.withDisplayName('NewAnalysisModal')(
         onDropAccepted: files => {
           const tool = getTool(files.pop().path)
           setCurrentTool(tool)
-          currentRuntime && !isRuntimeDeletable(currentRuntime) && currentRuntimeTool !== tool ? onSuccess() : enterNextViewMode(tool, newAnalysisMode)
+          currentRuntime && !isRuntimeDeletable(currentRuntime) && currentRuntimeTool !== tool ?
+            onSuccess() :
+            enterNextViewMode(tool, analysisMode)
           uploadFiles()
         }
       }, [() => div({
@@ -162,7 +191,10 @@ export const NewAnalysisModal = Utils.withDisplayName('NewAnalysisModal')(
           resetView()
           onSuccess()
           openUploader()
-        }, style: { marginTop: '1rem', fontSize: 16, lineHeight: '20px', ...Style.elements.card.container, alignItems: 'center', width: '100%', height: 150, backgroundColor: colors.dark(0.1), border: `1px dashed ${colors.dark(0.7)}`, boxShadow: 'none' }
+        }, style: {
+          marginTop: '1rem', fontSize: 16, lineHeight: '20px', ...Style.elements.card.container, alignItems: 'center', width: '100%', height: 150,
+          backgroundColor: colors.dark(0.1), border: `1px dashed ${colors.dark(0.7)}`, boxShadow: 'none'
+        }
       }, [
         div(['Or Click / Drag to upload an analysis file']),
         icon('upload-cloud', { size: 25, style: { opacity: 0.4, marginTop: '0.5rem' } })
@@ -213,12 +245,16 @@ export const NewAnalysisModal = Utils.withDisplayName('NewAnalysisModal')(
             options: ['python3', 'r']
           })
         ])]),
-        (isJupyter || toolLabel === tools.RStudio.label) && (currentRuntime && !isRuntimeDeletable(currentRuntime) && currentRuntimeTool !== toolLabel) && div({ style: { backgroundColor: colors.warning(0.1), margin: '0.5rem', padding: '1rem' } }, [
-          h(WarningTitle, { iconSize: 16 },
-            [span({ style: { fontWeight: 600 } }, ['Environment Creation'])]
-          ),
-          div({ style: { marginBottom: '0.5rem', marginTop: '1rem' } }, ['You have a non-deletable environment associated with another application.']),
-          div(['You may create an analysis, but must wait for your current environment to finish processing and get a suitable environment to run it.'])
+        (isJupyter || toolLabel === tools.RStudio.label) &&
+        (currentRuntime && !isRuntimeDeletable(currentRuntime) && currentRuntimeTool !== toolLabel) &&
+        div({ style: { backgroundColor: colors.warning(0.1), margin: '0.5rem', padding: '1rem' } }, [
+          h(WarningTitle, { iconSize: 16 }, [span({ style: { fontWeight: 600 } }, ['Environment Creation'])]),
+          div({ style: { marginBottom: '0.5rem', marginTop: '1rem' } }, [
+            'You have a non-deletable environment associated with another application.'
+          ]),
+          div([
+            'You may create an analysis, but must wait for your current environment to finish processing and get a suitable environment to run it.'
+          ])
         ]),
         div({ style: { display: 'flex', justifyContent: 'flex-end', marginTop: '1rem' } }, [
           h(ButtonPrimary, {
@@ -244,8 +280,8 @@ export const NewAnalysisModal = Utils.withDisplayName('NewAnalysisModal')(
     }
 
     const width = Utils.switchCase(viewMode,
-      [newEnvironmentMode, () => 675],
-      [newAnalysisMode, () => 450],
+      [environmentMode, () => 675],
+      [analysisMode, () => 450],
       [Utils.DEFAULT, () => 450]
     )
